extract feebox and dialog from issue
